requirements.txt
pandas
python-dotenv
pydantic

# For OpenAI
openai

# For MCP
fastmcp

# For API
fastapi